Multiple adjustments to JS file
Add new form submit dom element and adjust form dom element. Adjusted the object consturctor and changed it to a class. Changed the displayBook() function to by using a for loop instead of a forEach loop but maintained displaying book content in a card format. Fially, changed the event listener from the form "submit" option to using the submit button instead and declaring the book elements into the dom making the code much cleaner and more functional. Deleted the reset form function as well.